+++ src/main/java/hr/avrbanac/inanis/shaders/AbstractShaderProgram.java
@@ -3,18 +3,23 @@ package hr.avrbanac.inanis.shaders;
 import hr.avrbanac.inanis.InanisErrorCode;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.joml.Matrix4f;
+import org.joml.Vector3f;
+import org.lwjgl.BufferUtils;
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL20;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
+import java.nio.FloatBuffer;
 
 import static hr.avrbanac.inanis.Inanis.exitWithError;
 import static hr.avrbanac.inanis.InanisConfig.MAX_SHADER_LOG_SIZE;
 
 public abstract class AbstractShaderProgram {
     private static final Logger LOG = LogManager.getLogger(AbstractShaderProgram.class);
+    private static FloatBuffer matrixBuffer = BufferUtils.createFloatBuffer(16);
     private int programId;
     private int vertShaderId;
     private int fragShaderId;
@@ -29,6 +34,30 @@ public abstract class AbstractShaderProgram {
         bindAttributes();
         GL20.glLinkProgram(programId);
         GL20.glValidateProgram(programId);
+        getAllUniformLocations();
+    }
+
+    protected void loadFloat(int location, float value) {
+        GL20.glUniform1f(location, value);
+    }
+
+    protected void loadVector(int location, Vector3f vector) {
+        GL20.glUniform3f(location, vector.x, vector.y, vector.z);
+    }
+
+    protected void loadBoolean(int location, boolean value) {
+        GL20.glUniform1f(location, value ? 1f : 0f);
+    }
+
+    protected void loadMatrix(int location, Matrix4f matrix) {
+        matrix.get(matrixBuffer);
+        GL20.glUniformMatrix4fv(location, false, matrixBuffer);
+    }
+
+    protected abstract void getAllUniformLocations();
+
+    protected int getUniformLocation(String uniformName) {
+        return GL20.glGetUniformLocation(programId, uniformName);
     }
 
     public void start() {